requirements.txt
fastapi[all]
python-dotenv
google-generativeai
yfinance
pandas